Avoid the last clone in try_recv
This removes the clone mentioned #18 for try_recv if the last receiver reads a message that would move the queue forward.

Benchmark results are encuraging:

* `one_to_one` showed a 1% drop in performance
* `one_to_two` it showed a 9% improvement in performance

The results seem worth the change as we'll likely see much more significant improvements for more complex data structures that need to be cloned instead of copied.

Signed-off-by: Heinz N. Gies <heinz@licenser.net>