Add CapacityLimiter trait
A limiter tracks the contents of a channel and determines if sending an
element to the channel will block or cause an overflow.  This allows
implementing more complex limits such as a limit on the total memory
size of elements in this channel.  The ususal solution of using a
separate semaphore to manage such limits does not work with broadcast
channels as only the channel tracks when the last copy of a message is
removed.

The AdjustableCapacityLimiter trait is split out from the
CapacityLimiter to simplify implementations that do not want to support
adjusting their capacity.  This also allows more complex adjustments
than just replacing the limiter (see later patches for an example).

A default implementation (ElementCountLimiter and FullError) is provided
that matches the current API.  This is set as the default generic
parameters to minimize API breakage (and this patch alone does not cause
an API break).