Add SendError::Rejected
This avoids a deadlock when there is not enough room in the channel to
send a message even after the channel is empty.  This can happen if you
run `set_capacity(0)` on a normal (element-count-limited) channel, or on
other channels if `try_add` returns an error when the channel is empty.