`Sender::broadcast` now returns `Send<'static>`
The static lifetime allows the `Send` to be kept around in the same
place/type as `Sender` w/o running into self-reference issues.

Since the inner struct is in an `Arc`, cloning is cheap and this change
doesn't affect the performance of `Sender::broadcast`.

We could just drop the lifetime parameter entirely at this point but
let's keep it for backwards compatibility.

Fixes #39.